# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peer Rater Api Documentation",
        "description": "Peer Rater Api Documentation",
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "register the user to the database",
                "description": "Returns the authentication infomation",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user@admin.com"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "User name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "PeerRater"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "PeerRater"
                    },
                    {
                        "name": "isInstructor",
                        "in": "path",
                        "description": "User access level",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "user login",
                "description": "Returns the authentication infomation",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user@admin.com"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "PeerRater"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "object",
                                            "example": "eyJ0eXAiOiJK..."
                                        },
                                        "token_type": {
                                            "type": "object",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "object",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "user logout",
                "description": "Returns the authentication infomation",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get login user infomation",
                "description": "Returns the authentication infomation",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "object",
                                            "example": 11
                                        },
                                        "name": {
                                            "type": "object",
                                            "example": "tom"
                                        },
                                        "email": {
                                            "type": "object",
                                            "example": "tom@tom.ca"
                                        },
                                        "email_verified_at": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "isInstructor": {
                                            "type": "object",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "object",
                                            "example": "2022-04-04T02:29:04.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "object",
                                            "example": "2022-04-04T02:29:04.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Error, please check server configuration"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
}